Separated models into two different files
