+++ src/models/user.ts
@@ -12,7 +12,8 @@ import { FacilityInitiatedFacilityRelationshipRequest, FacilityInitiatedFacility
 import { GlobalAccessControl, GlobalAccessControlResponse } from './globalAccessControl'
 import { HeartRateCapturedDataPoint, HeartRateDataSet, HeartRateDataSetListResponse, HeartRateDataSetResponse, HeartRateDataSets, HeartRateDataSetSorting } from './heartRateDataSet'
 import { HeightMeasurement, HeightMeasurementData, HeightMeasurementListResponse, HeightMeasurementResponse, HeightMeasurements, HeightMeasurementSorting } from './heightMeasurement'
-import { MSeriesChallenge, MSeriesChallengeFocus, MSeriesChallengeListResponse, MSeriesChallengeParticipant, MSeriesChallengeParticipantResponse, MSeriesChallengeRelationship, MSeriesChallengeResponse, MSeriesChallenges, MSeriesChallengeSorting, MSeriesChallengeType } from './mSeriesChallenge'
+import { MSeriesChallenge, MSeriesChallengeFocus, MSeriesChallengeListResponse, MSeriesChallengeRelationship, MSeriesChallengeResponse, MSeriesChallenges, MSeriesChallengeSorting, MSeriesChallengeType } from './mSeriesChallenge'
+import { MSeriesChallengeParticipant, MSeriesChallengeParticipantResponse } from './mSeriesChallengeParticipant'
 import { MSeriesCapturedDataPoint, MSeriesDataSet, MSeriesDataSetListResponse, MSeriesDataSetResponse, MSeriesDataSets, MSeriesDataSetSorting } from './mSeriesDataSet'
 import { MSeriesFtpMeasurement, MSeriesFtpMeasurementListResponse, MSeriesFtpMeasurementResponse, MSeriesFtpMeasurements, MSeriesFtpMeasurementSorting } from './mSeriesFtpMeasurement'
 import { OAuthProviders, OAuthService, OAuthServiceData, OAuthServiceListResponse, OAuthServiceResponse, OAuthServices } from './oauthService'
